Biblatex-Support: recognize addbibresource - ID: 3315923
Patch applied from [Sourceforge issue][1].

When using biblatex the bib file is included by

    \addbibresource{library.bib}

Vim-Latex does not recognize this command and so the outline does not
work.

The attached tiny patch replaces the search functions in the texviewer.
It is a quite simple patch, just works for this special case and makes
latex suite recognize the bib file.

[1]: http://sourceforge.net/tracker/index.php?func=detail&aid=3315923&group_id=52322&atid=466458